requirements.txt
yfinance>=0.2.40
easyocr>=1.7.0
Pillow>=10.0.0
pytesseract>=0.3.10
tqdm>=4.65.0
